+++ semgrep-core/finding/files_finder.mli
@@ -1,2 +1,6 @@
+(*s: semgrep/finding/files_finder.mli *)
 
+(*s: signature [[Files_finder.files_of_dirs_or_files]] *)
 val files_of_dirs_or_files: string list -> Common.filename list
+(*e: signature [[Files_finder.files_of_dirs_or_files]] *)
+(*e: semgrep/finding/files_finder.mli *)